Fixed drawBmp for icons

+++ src/GfxUi.h
@@ -41,7 +41,7 @@ class GfxUi {
     void setTextColor(uint16_t c);
     void setTextColor(uint16_t c, uint16_t bg);
     void setTextSize(uint16_t x);
-    void drawBmp(String filename, uint8_t x, uint16_t y);
+    void drawBmp(String filename, uint16_t x, uint16_t y);
     void drawProgressBar(uint16_t x, uint16_t y, uint16_t w, uint16_t h, uint8_t percentage, uint16_t frameColor, uint16_t barColor);
 
   private:
@@ -55,4 +55,3 @@ class GfxUi {
 };
 
 #endif
-
